Rename local_plan_cache.[c_h] files to fast_path_plan_cache.[c_h]
