+++ src/backend/distributed/planner/fast_path_plan_cache.c
@@ -11,12 +11,12 @@
 
 #include "distributed/pg_version_constants.h"
 
-#include "distributed/listutils.h"
-#include "distributed/local_executor.h"
-#include "distributed/local_plan_cache.h"
 #include "distributed/deparse_shard_query.h"
+#include "distributed/fast_path_plan_cache.h"
 #include "distributed/citus_ruleutils.h"
 #include "distributed/insert_select_planner.h"
+#include "distributed/listutils.h"
+#include "distributed/local_executor.h"
 #include "distributed/metadata_cache.h"
 #include "distributed/multi_executor.h"
 #include "distributed/version_compat.h"
